Calculator Exmaples

+++ GreetingProject/src/com/batch3/day3/WhileLoopExmaple.java
@@ -6,7 +6,7 @@ public class WhileLoopExmaple {
 	public static void main(String[] args) {
 		System.out.print("Please input number between 1 to 100 :");
 		Scanner sc = new Scanner(System.in);
-		int rowNumber = sc.nextInt();
+		long rowNumber = sc.nextInt();
 		while(rowNumber>= 0 && rowNumber <= 100) {
 			System.out.println("Row Number "+rowNumber+" mark is "+(100-rowNumber));
 			// last line increasement or dec
